Main화면에서 BottomNavigationview 연동 및 다른 Fragment와 연동
