--- SPOTeam_android/app/src/main/java/com/example/spoteam_android/HomeFragment.kt
@@ -1,4 +1,0 @@
-package com.example.spoteam_android
-
-class HomeFragment {
-}
