--- SPOTeam_android/app/src/main/java/com/example/spoteam_android/StudyFragment.kt
@@ -1,18 +1,0 @@
-package com.example.spoteam_android
-
-import android.os.Bundle
-import androidx.fragment.app.Fragment
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-
-class StudyFragment : Fragment() {
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View? {
-        // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_study, container, false)
-    }
-}
