+++ SPOTeam_android/app/src/main/java/com/example/spoteam_android/ui/category/CategoryFragment.kt
@@ -1,24 +1,24 @@
-package com.example.spoteam_android.ui.home
+package com.example.spoteam_android.ui.category
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.spoteam_android.databinding.FragmentHomeBinding
+import com.example.spoteam_android.databinding.FragmentCategoryBinding
 
-class HomeFragment : Fragment() {
 
-    lateinit var binding: FragmentHomeBinding
+class CategoryFragment : Fragment() {
+
+    lateinit var binding: FragmentCategoryBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentHomeBinding.inflate(inflater, container, false)
+        binding = FragmentCategoryBinding.inflate(inflater, container, false)
 
         return binding.root
     }
-
 }
