--- SPOTeam_android/app/src/main/java/com/example/spoteam_android/ui/dashboard/DashboardFragment.kt
@@ -1,26 +1,0 @@
-package com.example.spoteam_android.ui.dashboard
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.fragment.app.Fragment
-import androidx.lifecycle.ViewModelProvider
-import com.example.spoteam_android.databinding.FragmentDashboardBinding
-
-class DashboardFragment : Fragment() {
-
-    lateinit var binding: FragmentDashboardBinding
-
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        binding = FragmentDashboardBinding.inflate(inflater, container, false)
-
-        return binding.root
-    }
-}
