--- SPOTeam_android/app/src/main/java/com/example/spoteam_android/ui/dashboard/DashboardViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.spoteam_android.ui.dashboard
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class DashboardViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is dashboard Fragment"
-    }
-    val text: LiveData<String> = _text
-}
