--- SPOTeam_android/app/src/main/java/com/example/spoteam_android/ui/home/HomeViewModel.kt
@@ -1,13 +1,0 @@
-package com.example.spoteam_android.ui.home
-
-import androidx.lifecycle.LiveData
-import androidx.lifecycle.MutableLiveData
-import androidx.lifecycle.ViewModel
-
-class HomeViewModel : ViewModel() {
-
-    private val _text = MutableLiveData<String>().apply {
-        value = "This is home Fragment"
-    }
-    val text: LiveData<String> = _text
-}
