+++ SPOTeam_android/app/src/main/java/com/example/spoteam_android/ui/home/community/CommunityHomeFragment.kt
@@ -1,4 +1,4 @@
-package com.example.spoteam_android.ui.home
+package com.example.spoteam_android.ui.home.community
 
 import android.os.Bundle
 import android.view.LayoutInflater
@@ -7,15 +7,15 @@ import android.view.ViewGroup
 import androidx.fragment.app.Fragment
 import com.example.spoteam_android.databinding.FragmentHomeBinding
 
-class HomeFragment : Fragment() {
+class CommunityHomeFragment : Fragment() {
 
     lateinit var binding: FragmentHomeBinding
-
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
+
         binding = FragmentHomeBinding.inflate(inflater, container, false)
 
         return binding.root
