+++ SPOTeam_android/app/src/main/java/com/example/spoteam_android/ui/jjim/BookmarkFragment.kt
@@ -1,24 +1,23 @@
-package com.example.spoteam_android.ui.home
+package com.example.spoteam_android.ui.mypage
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.spoteam_android.databinding.FragmentHomeBinding
+import com.example.spoteam_android.databinding.FragmentBookmarkBinding
 
-class HomeFragment : Fragment() {
+class BookmarkFragment : Fragment() {
 
-    lateinit var binding: FragmentHomeBinding
+    lateinit var binding: FragmentBookmarkBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentHomeBinding.inflate(inflater, container, false)
+        binding = FragmentBookmarkBinding.inflate(inflater, container, false)
 
         return binding.root
     }
-
 }
