+++ SPOTeam_android/app/src/main/java/com/example/spoteam_android/ui/mypage/MyPageFragment.kt
@@ -1,24 +1,23 @@
-package com.example.spoteam_android.ui.home
+package com.example.spoteam_android.ui.mypage
 
 import android.os.Bundle
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.spoteam_android.databinding.FragmentHomeBinding
+import com.example.spoteam_android.databinding.FragmentMypageBinding
 
-class HomeFragment : Fragment() {
+class MyPageFragment : Fragment() {
 
-    lateinit var binding: FragmentHomeBinding
+    lateinit var binding: FragmentMypageBinding
 
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View {
-        binding = FragmentHomeBinding.inflate(inflater, container, false)
+        binding = FragmentMypageBinding.inflate(inflater, container, false)
 
         return binding.root
     }
-
 }
