+++ SPOTeam_android/app/src/main/res/layout/fragment_category.xml
@@ -1,14 +1,19 @@
 <?xml version="1.0" encoding="utf-8"?>
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
+<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:gravity="center"
-    android:orientation="vertical">
+    android:layout_height="match_parent">
 
     <TextView
+        android:id="@+id/communityHome_title_tv"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:text="Category Fragment"
-        android:textSize="24sp"/>
-</LinearLayout>
+        android:layout_marginTop="20dp"
+        android:text="Category"
+        android:textSize="40sp"
+        app:layout_constraintEnd_toEndOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintTop_toTopOf="parent" />
 
+
+</androidx.constraintlayout.widget.ConstraintLayout>
