+++ SPOTeam_android/app/src/main/res/menu/bottom_nav_menu.xml
@@ -1,4 +1,5 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android">
+
     <item
         android:id="@+id/navigation_home"
         android:icon="@drawable/ic_home"
