+++ SPOTeam_android/app/src/main/res/navigation/mobile_navigation.xml
@@ -13,13 +13,13 @@
 
     <fragment
         android:id="@+id/navigation_dashboard"
-        android:name="com.example.spoteam_android.ui.dashboard.DashboardFragment"
+        android:name="com.example.spoteam_android.ui.category.CategoryFragment"
         android:label="@string/title_dashboard"
-        tools:layout="@layout/fragment_dashboard" />
+        tools:layout="@layout/fragment_category" />
 
     <fragment
         android:id="@+id/navigation_notifications"
-        android:name="com.example.spoteam_android.ui.notifications.NotificationsFragment"
+        android:name="com.example.spoteam_android.ui.mypage.MyPageFragment"
         android:label="@string/title_notifications"
-        tools:layout="@layout/fragment_notifications" />
+        tools:layout="@layout/fragment_mypage" />
 </navigation>
